Change log levels and remove unnecessary exception handling
Adjusted the log levels of loggerContext and overloadTimer. The level of loggerContext has been changed from ERROR to WARN, and the log level of overloadTimer has been updated from DEBUG to WARN. Also removed the printStackTrace call in the InterruptedException catch block of BurstRateLimiter, as it was contributing to unnecessary noise in the logs. Minor cleanups have also been made to remove redundant spaces and lines.